ui changes/code restructuring
